Consolidated duplicated code with updatePropertyForMem.
Added ability to modify the URL of the mem.

Added ability to debug the opengraph code.

Fixed issue where the mem that was returned back after editing didn't have the edits applied.

Added some more URL params to filter out.

Added a modified version of AutoresizeTextarea so we can forward some specific events back out of it.

Added a max width to the single mem layout.

Added vitest so that we can test the opengraph parse independently from everything else.